+++ src/transactions/transactions.module.ts
@@ -1,0 +1,14 @@
+import { Module } from '@nestjs/common';
+import { TypeOrmModule } from '@nestjs/typeorm';
+import { Expense } from '../expenses/entities/expense.entity';
+import { Income } from '../income/entities/income.entity';
+import { TransactionsController } from './transactions.controller';
+import { TransactionsService } from './transactions.service';
+
+@Module({
+  imports: [TypeOrmModule.forFeature([Expense, Income])],
+  controllers: [TransactionsController],
+  providers: [TransactionsService],
+  exports: [TransactionsService],
+})
+export class TransactionsModule {}